fix(styles): move LLMResponseParser to core so CLI paths need no PySide6
A PySide6-less --style-create previously died with a misleading
'No module named PySide6' because core modules imported the parser
from gui.llm_utils. The class now lives in core/llm_parsing.py; gui
re-exports it for existing dialog imports, and all six core import
sites (styles analyzer/applicator, prompt_enhancer_llm, layout
designer/prompt_helper) point at core. Issue #37.
